:recycle: Refactored Board, Game and Hole into their own component files
